+++ R/roxygen2.R
@@ -24,7 +24,7 @@ extract_diff_files <- function(files) {
 #' hook such as [roxygen2::roxygenize()] must run at all or not.
 #' @param root The root of project.
 #' @keywords internal
-extract_diff_root <- function(root = here::here()) {
+extract_diff_root <- function(root = ".") {
   assert_is_git_repo(root)
   repo <- git2r::repository(root)
   if (length(git2r::reflog(repo)) == 0) {
@@ -57,7 +57,7 @@ extract_diff_root <- function(root = here::here()) {
 #' diff_requires_run_roxygenize()
 #' }
 #' @export
-diff_requires_run_roxygenize <- function(root = here::here()) {
+diff_requires_run_roxygenize <- function(root = ".") {
   if (rlang::with_handlers(withr::with_namespace("git2r", FALSE), error = function(...) TRUE)) {
     generic <- paste0(
       " Please add the package as a dependency to ",