+++ inst/hooks/exported/codemeta-description-updated.R
@@ -1,5 +1,21 @@
 #!/usr/bin/env Rscript
 
+"A hook to make sure DESCRIPTION hasn’t been edited more recently than
+codemeta.json.
+
+Usage:
+  codemeta-description-updated [--root=<root_>] <files>...
+
+Options:
+  --root=<root_>  Path relative to the git root that contains the R package
+                  root [default: .].
+
+" -> doc
+
+
+arguments <- docopt::docopt(doc)
+setwd(arguments$root)
+
 # adapted from https://github.com/lorenzwalthert/precommit/blob/f4413cfe6282c84f7176160d06e1560860c8bd3d/inst/hooks/exported/readme-rmd-rendered
 if (!file.exists("DESCRIPTION")) {
   rlang::abort("No `DESCRIPTION` found in repository.")